(MAINT) Add cultivars and update front page
This commit adds the new cultivars page, which highlights a few projects
folks may want to purchase or back. This will be updated on a regular
basis to stay fresh with new releases/stuff to highlight currently.

It also updates the front page to point to the new section as well as
change the wording regarding the games and point to the open collective
for the cooperative.