(MAINT) Incorporate FF dev edits
This commit reworks the Flagrant Factions text to incorporate all of the
dev edits sourced in the google doc. From this point on, the source of
truth for the project is this project.

In this conversion, no additional links or shortcodes are used; as much
as possible, this update keeps the text in plain markdown. After all
editing is completed, we can convert to using shortcodes and make other
improvements to interactivity and referability.

This commit also:

+ standardizes on using `+` for entries in unordered lists
+ standardizes on using `1.` for entries in ordered lists
+ standardizes on using semantic linefeeds for easier editing