(MAINT) Fix link warnings
This commit updates the links in several blog posts to use the ref
shortcode with `{{% ... %}}` tags instead of `{{< ... >}}` tags to help
resolve warnings during build about invalid shortcodes. This is related
to how hugo's parsing works - the `%` delimited tags pass the actual
HTML they generate through the end template even when the surrounding
text - in this case, markdown - is processed by another template.

This commit also updates the link in the _Fever Black Mountain_ post to
point correctly at the location for _Flagrant Picaresque_.