(MAINT) Add minimal VS Code settings
Prior to this commit, the project had no VS Code settings configured.

This commit adds two new settings:

1. It sets the default width for reflow to 100 characters.
2. It adds rulers at 76, 80, 96, and 100 characters.

In both cases, this is to support writing text of an appropriate width
for easier editing and viewing. We may choose to extend the furthest
ruler to a wider length in the future (such as 120), but the ruler at
80 characters is the max-width for non-scrolling code blocks on the
rendered site.