(MAINT) Standardize frontmatter metadata
Prior to this change, the usage of frontmatter keys and formatting was
messy, inconsistent, and difficult to understand.

This change standardizes the metadata across the site and updates it
where appropriate.

Every file _must_ define:

- `title` - Used in the page metadata and stands as the first header for
  any page.
- `summary` - Used as the description in the metadata and for previewing
  the page in various circumstances, especially with the `section`
  shortcode and when paginating.

Every blog post must also define:

- `author` - the person who wrote the post
- `date` - the published date for the post
- `categories` - the post must fit under one or more categories

Game texts must define `date` as the day they are done with editing and
no longer considered WIP.

This change also sets the **enableGitInfo** setting to `true`, which
auto-populates the last modified date for any files with the date they
were last committed.