(MAINT) Remove unused language modes
Prior to this change, the site configuration declared support for
content written in Russian and Chinese, despite having no pages in
either language.

This change comments those entries out in the configuration and corrects
the site so that visitors won't be mislead about what languages the
content is currently available in.