(MAINT) Move menu entry definitions into config
Prior to this change, the configuration to put the blog posts and the
horticulturists pages in the site navigation menu lived in the
frontmatter metadata for those sections.

This change moves those definitions into the site configuration, making
that configuration file the only required configuration item for the
site nav (outside of the auto-generated entries for texts in the `games`
folder).