(TEMP) Override menu partial for relative entries
Prior to this change, entries in the site configuration's file for the
site navigation menu treated relative links and external links the same
way, always opening them in a new tab.

This change clobbers the `docs/menu-hugo.html` partial from the
**hugo-book** theme so that site-relative links in the menu may open in
the same tab, like all other site-relative links.

In the future, if the PR for this fix is accepted into **hugo-book**,
this change should be reverted.