(MAINT) Schematize data files
Prior to this change, the data files for Flagrant Factions used an
undefined schema. This change introduces schemas for every data type in
the existing files, updates the files to match the schema, and adds
association of the schemas to the files for VS Code validation.

Future commits should add more schemas, associate the data files with
Front Matter (to provide a UI for updating them), and ensure the schemas
and data files can be published and used in the site.