(MAINT) Add data types and folders to Front Matter
Prior to this change, the data files for the project needed to be
updated by hand with the available schemas.

With this change, again taking advantage of the preview feature for
splitting the config made available in estruyf/vscode-front-matter#412,
we now have the various Flagrant Factions data types and their data
entry files incorporated to Front Matter.

This will allow us to modify those values in a nice UI instead of
manually in YAML.