(MAINT) Schematize the Hugo Configuration
This change adds schemas for Hugo's configuration file to the project
and associates them with the `config` folder in the project root. The
`config` folder allows a site maintainer to split their configuration
across multiple files and define different settings for their
environments as subfolders.

- `base.yaml` applies the full hugo configuration as the schema to the
  `config.yaml` file in the project root.
- `book.yaml` applies the schema for the hugo-book theme's parameters
  to any file named `params.yaml` in the `config` folder or its
  children.
- `sharing.yaml` applies the schema for hugo's parameters that affect
  the Open Graph and Twitter Card internal shortcodes to any file named
  `params.yaml` in the `config` folder or its children.
- `toroidal.yaml` applies the schema for the Toroidal theme's parameters
  to any file named `params.Toroidal.yaml` in the `config` folder or its
  children.

The reason that `book.yaml` applies to `params.yaml` instead of
`params.Book.yaml` is that the naming convention for split configs is
that the name represents the dotpath for the configuration items in that
file. So `params.Toroidal.yaml` maps its settings to the `Toroidal` key
in the `params` blob of the composed configuration.

Because hugo-book's parameters are not namespaced within an object, they
need to be placed in `params.yaml` or in `config.yaml` under the
`params` blob. Luckily, we can associate multiple schemas with a single
file type and that works fine.

The model for this approach then is:

1. Site-wide basic configuration goes in `config.yaml`
1. Site-wide sharing parameters and hugo-book parameters go in
   `config/params.yaml`.
1. Toroidal parameters go in `config/params.Toroidal.yaml`.

While this schematization is base-level useful for authoring and
maintaining configurations, it's not a delightful UX. For that, we will
need to use Front Matter to treat the configuration files as data files,
which will give us a friendlier UX for managing them.