(MAINT) Use split hugo configuration
Prior to this change, the hugo configuration was defined in the
`config.yaml` file at the project root. This change takes advantage of
hugo's [configuration directory][01] to migrate the site configuration
into the `config` folder in the project root.

At the same time, it moves the site parameters from the **params** key
in the base configuration into the `params.yaml` file.

This change takes advantage of the schematized configuration files and
makes reasoning about and managing the configuration a little easier.
It has the nice byproduct of introducing an easier way for users to
override configurations on a per-environment basis and for splitting out
more of the configuration in the future, if needed.

[01]: https://gohugo.io/getting-started/configuration/#configuration-directory