(MAINT) Add Hugo paramter configs to Front Matter
Prior to this change, the configuration files for Hugo were scehamtized,
which made editing them directly in YAML more friendly.

This change converts those defined schemas to Front Matter's data type
and file definitions, allowing us to configure Hugo in a friendly UI.

This change also renames the data types for Flagrant Factions to better
distinguish between data types by purpose, since the data types can't be
placed in subfolders yet.

This change doesn't add a UI for the base Hugo configuration file. At
this time, that configuration is too large and we don't have sufficient
tools to use it. In the future, if estruyf/vscode-front-matter#397 is
implemented, we could simply reference the published schema.

In the meantime, we need to work on scripting to convert from the YAML
schema definition to a data type definition that Front Matter will
accept. Ideally, we'll build a workflow that will allow us to author in
YAML and automatically convert/export to JSON Schema and a Front Matter
definition for easier usage.